Fix CXO-5000: Remaining tests: toTimeSeries (both mocked and integration tests)

+++ coko/com.elster.jupiter.calendar/src/main/java/com/elster/jupiter/calendar/CalendarService.java
@@ -96,14 +96,14 @@ public interface CalendarService {
         EventSet add();
 
         interface EventBuilder {
-            EventSetBuilder withCode(int code);
+            EventSetBuilder withCode(long code);
         }
     }
 
     @ProviderType
     interface DayTypeBuilder {
         DayTypeEventOccurrenceBuilder event(String eventName);
-        DayTypeEventOccurrenceBuilder eventWithCode(int code);
+        DayTypeEventOccurrenceBuilder eventWithCode(long code);
         CalendarBuilder add();
     }
